all itenaries and filted 5 done

+++ src/Pages/Itinerary.js
@@ -23,10 +23,9 @@ const Itinerary = () => {
 					<h3 className="font-bold text-2xl my-2">{details.title}</h3>
 					<hr />
 					<p className="font-normal my-2">
-						<b>Day</b> :
-						{new Date(details.date).toLocaleDateString("en-US", {
-							weekday: "short",
-						})}
+						<b>Day</b> : {new Date(details.date).toLocaleDateString("en-US", {
+								weekday: "long",
+							})}
 						<br />
 						<b>Time</b> : {details.time}
 						<br />
